rename: WordList → WordListItem
- WordList → WordListItemDesktop
- WordListCompact → WordListItemMoblie
