+++ frontend/src/components/WordListItem/WordListItemDesktop.module.scss
