+++ frontend/src/components/WordListItem/WordListItemMobile.module.scss
