+++ frontend/src/components/WordListItem/types.ts
@@ -1,4 +1,4 @@
-export interface WordListProps {
+export interface WordListItemProps {
   KoreanWord: string;
   ForeignWord: string;
   isExpanded?: boolean;
